updated frontend wrapper logic to be more dynamic via JavaScript
- removed red border; fixes #6

Reason for this change: shown wrapper is now less likely to destroy the given
html-output from a styling point of view. Before, it was a div that's wrapped
around the inner plugins - now it is an extra element that'll be placed above
the next plugin-html.
Known: may also have some restrictions regarding styling, for example if the
next html-element isn't shown on a screen size.